fix: nullable accel/nullable close/lag rounding from CI compile report

+++ app/src/main/java/com/example/engine/TurboAnalyzer.kt
@@ -104,7 +104,7 @@ class TurboAnalyzer(
             overBoostEvents = overBoostCount,
             tipInCount = tipIns.size,
             lastLagMs = tipIns.lastOrNull()?.lagMs,
-            averageLagMs = if (lags.isEmpty()) null else lags.average(),
+            averageLagMs = if (lags.isEmpty()) null else lags.average().toLong(),
             chargeTempC = chargeTemp,
             wastegatePct = wastegate,
             boostVsRpm = boostByRpmBin.entries
